+++ test/src/callbacks.rs
@@ -3,15 +3,15 @@ use ocaml::{Error, Value};
 #[ocaml::func]
 #[ocaml::sig("('a -> 'b) -> 'a -> 'b")]
 pub unsafe fn apply1(f: Value, x: Value) -> Result<Value, Error> {
-    f.call(gc, x)
+    f.call1(gc, x)
 }
 
 #[ocaml::func]
 #[ocaml::sig("('a -> 'b) -> 'a -> 'b")]
 pub unsafe fn apply3(f: Value, x: Value) -> Result<Value, Error> {
-    let a = f.call(gc, x)?;
-    let b = f.call(gc, a)?;
-    f.call(gc, b)
+    let a = f.call1(gc, x)?;
+    let b = f.call1(gc, a)?;
+    f.call1(gc, b)
 }
 
 #[ocaml::func]
@@ -22,5 +22,5 @@ pub unsafe fn apply_range(f: Value, start: ocaml::Int, stop: ocaml::Int) -> Resu
         l = l.add(gc, stop - 1 - i)
     }
 
-    f.call(gc, l)
+    f.call1(gc, l)
 }