+++ CHANGES.md
@@ -8,6 +8,8 @@
 - Added `ocaml::import!` macro for calling OCaml functions from Rust
 - Added `ocaml::sig` proc-macro for generating `external` and type signatures
 - Added ocaml-build crate for generating OCaml code from `ocaml::sig` macros
+- Renamed `Value::call` to `Value::call1` and rewrote `Value::call` to take a variable number of
+  arguments
 
 ## 0.22.4